requirements.txt
python-multipart
python-jose[cryptography]
fastapi
fastapi-rss
pydantic
uvicorn[standard]
gunicorn  # Used as process manager for uvicorn processes

elasticsearch
argon2-cffi
requests

python-dotenv
CouchDB
pathvalidate
